Make Apps and Units use common instances
Per review, a given application or unit in the model should be
represented by the same instance no matter where it's referenced. This
necessitates reversing the relationship between Applications and Units,
since a given unit always has an unambiguous Application instance it can
refer to, but the list of Units visible for a given Application will
depend on the context in which it's being used.