Persist stored state before framework commit
Currently stored state is persisted only on when __setattr__ is called
on BoundStoredState which means any operations on data structures like
sets, lists or dicts are not persisted unless something like the
following is done:

self.state.myset = self.state.myset

Moreover, calling save_snapshot on every __setattr__ is expensive and
semantically incorrect.

In order to notify BoundStoredState objects about a commit that is about
to happen a `commit` event is defined on the framework itself and
BoundStoredState objects subscribe StoredStateData to it in the __init__
method. This is not done in StoredStateData's __init__ as __init__ is
not called when an object is unpickled.

A pre-commit event is defined in order to allow any subclasses of Object
to persist their state before the commit event is emitted.

Framework events need to have a correct handle which includes the
Framework type, e.g. "Framework/on/pre_commit", as opposed to having
something like "on/pre_commit".

In order to achieve that without special-casing the Framework class,
this change makes it a subclass of Object.

Since pickled data may contain NULL bytes we need to make sure that the
column type is semantically correct - so the snapshot data column is
changed to use the BLOB type:

https://www.sqlite.org/c3ref/bind_blob.html
"The result of expressions involving strings with embedded NULs is
undefined"