Use (pre)commit events for snapshoting state (#21)
Instead of saving the state on every change, the stored state
mechanism will now snapshot it right before a commit is made.
That's done by introducing two new events: pre_commit and
commit. The former allows touching the state right before a
commit, and the latter actually snapshots it for saving.