Refactor params to CharmBase constructor
Remove all arguments to the CharmBase constructor aside from the
framework to minimize the factors that charm devs have to deal with when
subclassing CharmBase.

This also adds a CharmEnv class to capture and hold all of the info from
the environment variables provided by Juju and other info from the charm
hook environment.