Implement dirty flag for StoredStateData
The Framework needs to trigger save_snapshot for StoredStateData only
when the state has actually changed.

The change also extends testing of nested data types to make sure
save_snapshot is called after changes to nested data types.

In order to avoid a separate loop for testing the dirty state
functionality it was merged into the one that tests modification
operations.

Operations are executed and validated multiple times:

* before the first commit to make sure that the modifications are
  available for use before any commit happens;
* after the first commit, to make sure that the assignment operation
  itself is not the only reason for the stored state to be marked as
  dirty.

"a" operand is obtained via a lambda as adding data to StoredStateData
does not result in full copies of the original data structures which
leads to side effects when a modification operation is executed multiple
times.