Add status property to units and applications
In order to allow status assignment for units and applications a status
descriptor is used.

Status for remote units and applications is set to unknown. While it is
possible to retrieve remote unit status via goal-state (but not
remote application status), this change does not aim to address this
case.