Invert entry-point semantics
By making the user code the target of the symlinks and relying on them
to call into the framework (and explicitly passing in the charm class),
we can eliminate the need to do path manipulation or require specific
file and class names. It also frees us up to package the framework as a
proper Python package and use standard dependency management tools.