ModelCache: Separate out a cache key and init args
Currently args to the get method of ModelCache are used both for lookup
in the cache and for passing arguments to `__init__`. This change
introduces a separate type called CacheKey in order to make it clear
which arguments will be used for lookup versus the ones that will be
passed into `__init__`.