Expose unit and app locality (#43)
Units or Applications can be local or remote which is currently not
exposed via properties of those types. This information is useful for
developers framework users and also for the framework itself in the
upcoming work.

For example, calling is_leader for a non-local unit will not be correct
and is_local will help to check for that condition.