Run python-black over the source tree.
Mostly this just wraps long lines and changes '' strings into ""
strings. The big reason to do this is that we can commit to it like
people do for "go fmt". It is the "One True Formatting" that means we
don't have to worry about it, and don't have to end up fixing flake8
warnings. You can configure an editor to automatically run black on your
source file, and that will allow you to not have to worry about
formatting again.