Change is_local to is_our and change visibility
Per one of the on-call discussions we agreed that exposing is_local
publicly as a property is more vague compared to using something like
this:

self.framework.model.unit == u

For use in the Unit and Application types themselves we can use
"private" variables.

The meaning of "local" is subjective so we decided to give "our" a try
for a private variable name.