Remove is-leader checking from status tests
Considering is-leader results will likely be cached in the future for
the lease time, tests that validate is-leader invocation presence will
break. In order to avoid that we can just check whether the required
status command is included or not.

If any extra hook tools will be called except for status-set and
is-leader, this will lead to an exception anyway because fakes for those
tools are not provided.