Use argv[0] to determine charm code relative link
A symlink to the charm code can be determined using argv[0], however,
there are some cross-platform aspects to take care of.

Moreover, Python puts an absolute path after symlink resolution into
argv[0] on Windows which makes the current method of determining a Juju
event name invalid on Windows when symlinks to a single entrypoint are
used. This can be addressed in Juju as a part of LP: #1854505 or by
using JUJU_CONTEXT_ID which contains an event name (but its format is
not publicly documented).