Address a review comment
As noted in

https://github.com/canonical/operator/pull/63#discussion_r353888507

the test case needs to be more clear about the initial conditions as we
now have several different test cases with different initial event
symlinks that must be present.

This change adds a dedicated step before any events are simulated that
validates the initial events. Initial events are marked as such in their
spec.