When emitting events, check for stale observers.
We can't hook into WeakValueDictionary references, as they are already
using a callback to cleanup the WeakValueDictionary. Rather than
implementing our own dictionary, piggy back off the fact that we already
iterate all the observers on each emit(). All we add in overhead is a
O(1) lookup to see if that path still exists.

If this becomes an issue, or we rework how we track observers so we
don't have to iterate the full list each time, we can look at
implementing a custom WeakValueDictionary that also triggers a cleanup
of the associated dict.