Raise an error if a minion calls set_podspec
While there is only one operator pod per application in CAAS models,
it periodically executes hooks for different units which may or may not
be leaders.

Charm developers should make sure that their unit is a leader before
trying to set a pod spec while the framework needs to check for
leadership early and raise an appropriate exception.