Cleanup observers before we add a new observer.
If we don't then there is a race condition where an observer which is
dying and an observer being registered at the same path cause either the
new observer to be removed, or the new observation to be registered
twice.
By cleaning first, we avoid double events. But we may need a different
solution for double events anyway.