Make object paths unique in a given framework.
This tracks all known objects for a given framework, and ensures that we
don't have multiple alive objects at the same path. There is still a bit
of a question around load_snapshot, as it is actively used to create a
pristine object, and test_mutable_types is failing because it is
actively creating a StoredStateData that mimics an existing state.