Remove a check for min-juju-version
Doing a version check to make sure functions are not silently skipped is
a good idea, however, with the currently implemented check for
min-juju-version, it becomes a requirement in every charm out there
developed with this framework. Having min-juju-version in every charm
that uses functions will become irrelevant with time as Juju versions
grow but it will still remain a step for a developer to do. It might be
better to avoid that.