Determine functions type only based on meta file
Juju does not remove charm directory entries that were not present in
the charm manifest on upgrade. For example, if a charm creates "actions"
directory during its operation after deployment, it will not be removed
on upgrade which will result in an exception with the current code.

The framework should not error out based on the presence of both
"actions" and "functions" directories, however, it needs to error out if
both functions.yaml and actions.yaml files are present. The presence of
one file or the other should determine the target directory to be used
("actions" or "functions").