Raise RelationNotFoundError in network-get
The semantics of network-get operation are similar to relation tools if
a relation id is provided explicitly.

Raising RelationNotFoundError can be useful in a higher-level code which
can try to obtain network info for an endpoint associated with a
specific relation id and then just use a relation name if there is no
such relation id (e.g. a relation does not exist anymore).