Add network endpoints to the model
network-get can be used even in the absence of relation ids to retrieve
bindings of (endpoint) names to spaces. Endpoint names can be relation
names or names from extra-bindings section of metadata.yaml.

While there is no functionality in Juju as of 2.7.0 to directly support
multiple bindings per endpoint, for cross-model relations the logic of
populating an ingress-address depends on a relation ID:
https://github.com/juju/juju/blob/juju-2.6.7/state/relationunit.go#L634-L658

Thus, EndpointMapping has the following types of keys:

* relation names (str);
* endpoint names from extra-bindings (str);
* Relation objects.

Dead relations are handled gracefully - if network-get fails due to a
dead relation, it is retried without specifying a relation-id.