Rename endpoints to bindings
* juju/juju/worker/uniter/runner/jujuc/network-get.go calls names
  supplied to network-get as "binding names";
* juju/charm/bundledata.go verifyEndpointBindings calls all relation
  endpoints and keys in extra-bindings "endpoints".

This change puts forward the term "Binding" as an object that has a name
and possibly a relation-id and associates it with network info
associated with a network space. Using the term "Endpoint" can be
confused with relation endpoints in a different context and with names
without a specific binding.