Split charm and functions meta
Per a review discussion, the framework should keep metadata objects
for charm and functions metadata separate in the same way Juju does it.

Moreover, the logic of placing a functions_type key into the metadata
was decided to be replaced so this change introduces the functions_dir
property on the FunctionsMeta object.

File loading functions are also moved to metadata objects themselves to
avoid spreading the related code around different modules.