Add support for functions (#79)
Hook tools are not possible to use for deferred function events as their
context is lost after the original execution. Functions were designed to
be one-time executions.

As a part of that, we also need to make sure that emitting random
function events without the necessary context is forbidden as well.