WIP Modify get_binding and add networking types
Per our discussions, we would like to provide a minimal public API for
now to have more freedom when it comes to possible backend changes in
Juju.

network-get includes a binding name as a mandatory parameter even in the
presence of a relation id which requires knowing a binding name at
a Binding object creation time which makes it more difficult to
implement lazy-loading. Likewise, there is no easy way to query a
relation name by a relation ID.