Makefile, test: avoid using backslash to protect quotes (#137)
Python has a rich enough variety of quotes that you should never need
to use a backslash to protect a single quote in a string; it a silly
little thing but it hurts readability. So this fixes that, and adds a
linter to make sure it stays fixed.

Modulo bugs in the linter of course.