Rework DB lock handling (#141)
Rework handling of exclusive db access

* sqlite3 Python module provides an implicit transaction management
  functionality when isolation_level argument is not None when passed to
  a connect method;
* sqlite3 itself by default operates in the autocommit mode which can
  be disabled by executing the "BEGIN" SQL command;
  https://www.sqlite.org/lockingv3.html#transaction_control
* BEGIN without EXCLUSIVE does not obtain an exclusive lock for the
  database and the default behavior is to run a transaction as DEFERRED;
  https://www.sqlite.org/lang_transaction.html#immediate

The desired behavior for the framework is holding an exclusive access to
the DB while a Framework object exists and is not closed which requires
maintaining exclusive access across transactions. This is possible to
achieve by using `PRAGMA locking_mode=EXCLUSIVE`.

With this change, it will not be possible to have two Framework objects
with exclusive access to the backend storage.