Handle empty CIDR field in network-get output
CIDR may be empty on some providers although it would seem that it can
be easily retrieved from local host interfaces. See LP: #1864102.

In order to avoid introducing None to charmers and let them work with
IPv{4,6}Network in any case an address itself can be transformed into a
valid /32 or /128 "network".