Disallow Bindings without relation ids
It is difficult to enforce the usage of Bindings associated with
relations vs the ones that are not. It is important to use relation ids
when working with network-get because its results are affected by them
(ingress-addresses and egress-subnets values depend on a relation ID).

To avoid situations where charms are written to support non-CMR
relations and use relation_id-less Bindings and work incorrectly with
CMR relations, it was decided to remove the support for them completely.

With this change charms will be enforced to handle the presence or lack
of a relation before using Binding objects.

When it comes to connecting to a deployed service before it is related
to anything, peer relations can be used as a peer relation is available
even if there is only one unit present. A relation id of a peer relation
can be used instead of an extra-binding or relation endpoint name for
which there are no relations yet. There can be multiple peer relation
endpoints defined so having multiple bindings is not a problem either.