Fixes for the test suite on Mac OSX. (#147)
On Mac, $TMP isn't '/tmp' so don't assert that the path exactly matches
'/tmp'. I don't think we care about the path as much as we care about
the filename. Also use assertRegex so that when the regex fails, it
tells you the string that it was matching against.

Also on Mac, time.monotonic() is more the 'time since the process
started' and not sychronized to some global epoch. The code assumed that
'0' was always going to be >30s away from time.monotonic() which isn't
true.