Enable stderr logging if JUJU_DEBUG is set.
When running under 'juju debug-hooks' JUJU_DEBUG is set, which we can
use to detect that we should also print to stderr, so that people
interactively running the charm can see what is going on. It also
changes the default log level to DEBUG (though juju-log is still at
INFO).