Add interface tls-certificates (requires)
This change introduces an implementation of the requires side of
tls-certificates which allows charms to use the CAClient object to
expose inputs for certificate generation by a certificate authority.

Currently easyrsa and Vault charms provide this interface while a client
charm can be any charm that needs to set up TLS termination for its
services.

Certificates and keys in ca_client_test_data.yaml were generated once
for the purposes of adding unit tests and are not used in any of the
running systems.