ops/main.py: support dispatch
starting with 2.8 juju will support having a single 'dispatch' binary
at the top level of the charm, isntead of (or as well as) supporting
individual binaries from charms or actions.

With this change the operator framework supports this, making the need
for symlinks from `hooks` go away. This should make things a lot
nicer.