Use an enum for load-balancing algorithm
Given that this interface is not specific to a particular load-balancer,
it needs to define its own set load-balancing algorithm strings which
can then be validated by the actual load-balancer implementation to see
if that algorithm is supported.

Backends can request that a specific load-balancing algorithm needs to
be used based on their implementation (possibly driven via charm
config).