ops/framework.py: remove StoredStateChanged. (#227)
The implementation was buggy. We can either fix it (pull #226), or we
can get rid of it.

We've had a general feeling that Charm.state should become Charm._stored
because other attributes shouldn't be peeking at it. And waiting for an
event indicating the state has changed is very much peeking at it.

Nothing inside our own framework depends on it, so for now we should just remove it.