ops/testing.py: Don't emit events for our data (#230)
Use the pattern of Harness.update_relation_data() to set all relation data, whether it is for our unit/app or for other units/apps. However, when updating our own relation data, there shouldn't be an associated relation-changed event. (Caveat an application data change for a peer relation when we aren't the leader.)

Remove the checks that the Harness wasn't letting you update_relation_data on your own data buckets.