test/test_model.py: Test Model using Harness._get_backend calls (#264)
Rework the test_model.py tests using the Harness instead of fake_script. This helps them be a bit clearer as to what they are actually testing.
This changes Model tests so it only tests how we interact with ModelBackend, rather than all the way to disk. But we should have reasonably coverage of ModelBackend and how it spawns commands.

This adds `Harness._get_backend_calls`, but that shouldn't be used by charmers.