ops.framework: Remove default target for Framework.observe
Our experience so far has shown that the implicit dependencies for
Framework.observe push people down a path that actually makes their
charms less maintainable. Especially when it comes to writing
components, it is better to have clearly named functions and clearly
defined targets.  For components, it is usually better to have private
methods, because the handlers are an implementation detail rather than
something you want to advertise to people who want to interact with
your component.

This is a breaking change (all charms that have been using
self.framework.observe(self.on.config_changed, self) will break. So we
want to make this change before we do an official release version.

fixes: #297