ops library support (#244)
With this change, an author wanting to export an interface or other component for use in the operator framework can easily do so, and a charm developer wanting to consume that can use that component with some minimal sanity checks around the API and provenance of the component they're using. Like a namespace package, but with some checks around versioning.
This does not get bytes on the disk, which is a separate problem; this is a way of having namespace with some more features.

For example, if the component is a standalone client for a particular interface, the code might be laid out like

    some_client/
    ├── README.md
    ├── requirements.txt
    ├── setup.py
    ├── some
    │   └── opslib
    │       ├── devel
    │       │   └── __init__.py
    │       └── stable
    │           └── __init__.py
    └── tests
        └── test_some.py

and then the charm developer that wanted to use this, once they have installed it (`pip install` or whatever is appropriate for their charm), they can use `ops.lib.use` to request the right API of the interface, specifying also the author to ensure they're getting the one they mean.

Some more changes are coming, but everything we have planned is backwards-compatible at this point (to the best of our knowledge, of course).