test/test_storage.py: Test formatting of stored values.
This shows how we store various types, mostly using the '!!' syntax as a
way to be explicit about things like python/tuple and sets.
Interestingly it stores the set as a dict mapping to null rather than a
list.
It does round trip most types.
Using '.strip()' makes it a little bit nicer. (any '\n' in the inner
yaml has to be expanded to a double '\n' in the outer yaml.)