ops/storage.py: Use SafeLoader.
SafeLoader supports almost everything we want except tuple, so we use
our own SimpleLoader to bring 'python/tuple' back in.
That still will not load custom types, etc.

We also pass the magic flags for style and flow to make the actual data
interchange a bit nicer if you ever have to introspect it.