ops: get version from git (#319)
This rejuggles how we get the version info.

When running from git, the version will be retrieved via the
equivalent of

    git describe --tags --long --dirty | sed -e 's/-/+/; y/-/./'

This produces a PEP440-compatible version from the semantic versioning
we're using and tagging, plus the git describe output. Meaning that if at
any point we need to compare versions, `packaging.version.parse` will do
the right thing.

Then, `setup.py` replaces the code that does that with a constant string
just before running `setup`, and swaps it back out after. The constant
version it replaces it with is the one it gets from the code it swapped out,
sneakily.

This means that people using ops from git will always report something we
can use to figure out what they're running (or as close as we can get to
that), while people running from a release will have no overhead.

This does imply a small change in the release workflow to stop github from
publishing broken tarballs, but that should be fine (the tarballs it's been
auto-producing so far have not been completely up to par so we needed to do
something there anyway). Another small change to the workflow is that 
the fallback version in `ops/version.py` needs to be bumped post release
(as opposed to bumped pre-release; this is a better order of things imho).