model: don't use --app for relation-get/-set before 2.7.1
Juju before 2.7.1 didn't support the `--app` parameter for
`relation-get` and `relation-set`, meaning that the current code
breaks in 2.7.0.

This fixes that by avoiding `--app` entirely when not needed,
returning an empty dict when needed for `relation-get` in an old Juju,
and raising a RuntimeError for `relation-set` in the same situation.

fixes: #223