many: move Model.name to be read from _ModelBackend
Unit name was already being read by _ModelBackend, as well as being
read by main(). We wanted to eliminate the duplication. We could either
pass more names down the stack, or we can pull them out of the backend.
This ended up a bit cleaner because we pass the Backend into Model
anyway, so it means having fewer args to the same object.

This also gave me an opportunity to eliminate a bunch of code that was
poking at os.environ just to set JUJU_UNIT_NAME.