test/test_main: move actions.yaml into the charm (#330)
The test suite had been setting up actions.yaml for each test that
needed it, but we might as well just set it up always. This makes the
test suite a bit less special, and will make it easier to use
'test_main' as a real deployable charm.

Name the tempdir something obvious, and cleanup the parent TMPDIR not
just the CHARMDIR.
Also use the tempdir for putting our state file, since we don't really
need yet another temp location.