test/test_model.py: racing tests
This is the general problem I've had with Travis. The tests passed on my
branch, and they passed on master. But chipaca landed a patch which
added calls to 'os' and I landed a patch that removed 'import os'. We
can live with occasional failures like this, or we can add the policy
that you can't land code unless it already includes the tip of master.