ops/framework.py: Handle events that are never observed.
When emitting an event, the first thing we do is save it, before we
build the list of notices that are observing it. We then trigger those
notices to be evaluated. Only if there is an observer that decides not
to defer, then we delete the event.
This updates the logic so that if there are no observers, we also delete
the event.
Because this bug has persisted for a while, also add a
Framework.remove_unreferenced_events which can find these orphaned
events and remove them from the database. Note that this isn't called by
default, but charms that know this has been an issue can add it to their
upgrade_charm handler.

fixes: 333