ops/framework.py: Don't save events if there are no observers.
A reasonably small optimization. If there are no observers, don't bother
to save the event in the first place.
It still saves the event before any notices are saved, so we can't have
a formatting problem that would cause the notices to be broken, it just
avoids a save + delete in the case that nothing is observing.