Unobserved events (#334)
When emitting an event, wait until we see an observer before we save it, then
build the list of notices that are observing it. Then trigger those
notices to be evaluated. If none of the observers that decide to defer,
then delete the event. (The old logic always saved the event, and if there were
no observers, would never delete the event, causing cruft in the database to grow.)

Because this bug has persisted for a while, also add a
Framework.remove_unreferenced_events which can find these orphaned
events and remove them from the database. Note that this isn't called by
default, but charms that know this has been an issue can add it to their
upgrade_charm handler.