test/test_lib.py: Add tests for new regex
Test that we support author with hyphens and LIBNAME= without the extra
whitespace.

Note that we could test that 'name' allows hyphens, but ops.lib.use()
itself uses another regex that refuses names that have dashes in them.
We probably should warn/error in parse if we see one like that.