test/test_main.py: Rewrite charm to use StoredState
(#363)

Rather than a custom database with a pickle, use the actual StoredState
class to send data back and forth between the test charm and the unit
test suite. We have to send strings instead of types.
This also allows us to clean up all the extra 'charm_config' attributes.
We also always define actions that can be run.