version-check to determine dispatch support
This branch introudces a change, and a simplification (or at least a
separation of concerns) for how the operator framework determines that
a juju is dispatch-aware, and how it determines what hook juju is
asking us to run.

In the existing code we used sys.argv[0] to determine the hook to run
unless JUJU_DISPATCH_PATH was set and a dispatch binary existed, in
which case that environment variable was used. This leads to
charmcraft setting this environment variable to signal what hook to
run (because it will be indirecting the call to the charm code to set
up PYTHONPATH etc). Together with it always creating a dispatch
binary, this means operator always thought it was running in a
dispatch-aware juju, which meant it never created all the missing
symlinks, which meant it never got events beyond those created by
charmcraft itself on non-dispatch-aware jujus.

With this branch, instead JUJU_VERSION is looked at to determine
dispatch-awareness, and JUJU_DISPATCH_PATH is used instead of argv[0]
to know which hook is wanted, and this separation means
non-dispatch-aware jujus should work with dispatch-enabled charms such
as what charmcraft produces.

One last change introduced is that, when creating the symlinks for
non-dispatch-aware jujus (or non-dispatch-enabled charms), the target
of the symlink will be the hook that is currently running, rather than
the charm entry point. This is because the hooks, like dispatch, may
well be shims that set up the environment (etc), without which the
charm entry point won't run.

fixes: #362