use controller-side storage automatically in some cases
Before this change, controller-side needed to be asked for explicitly.
However, in some cases we _know_ you should probably be using it: When
running in k8s and on a new enough Juju you almost certainly want to
enable it so you don't need to have persistent storage for your
controller pod.

This change makes it so that when we detect that the charm is in this
exact scenario, it uses controller-side automatically unless
explicitly requested not to.