fix up logging tests impacting global state so much
The way we were doing the log tests, and some main tests, meant that
we'd the logging module's global state was leaking, accumulating
handlers and otherwise messing things up. In some situations this
could lead to failures where logassert and our juju-log handling got
into a squabble due to arbitrary ordering of handlers.

This change means that we only actually set up our juju-log handler
when we need it, and actually reset the logging global state when we
do.

HTH, HAND.