Add config alias to CharmBase
As discussed on [Discourse][], using `self.model.config` can cause the
charm author to believe they are accessing the model-level config. It's
also extra typing for something that is going to be frequently accessed,
so like the other frequently accessed attributes like `self.app`,
`self.unit`, etc, a `self.config` alias seems easier and less likely to
lead to confusion.

[Discourse]: https://discourse.juju.is/t/first-steps-with-the-operator-framework/3006/12