give StoredFoo classes a useful __repr__ (#447)
Before this change the three StoredFoo classes (`StoredDict`,
`StoredList` and `StoredSet`) did not have a `__repr__`, meaning that
if for example you wrote them out to the log them you'd get e.g.

    <ops.framework.StoredList object at 0x7f8b3dca0b50>

which is not particularly useful. With this change instead you'd get

    ops.framework.StoredList(["hi"])

which is not completely correct (in that you can't `eval(repr(obj))`
to get back `obj`), but it is at least more useful than what was there
before.