cache the backend goal-state call
This works exactly the same way as the backend.is_leader method. It is unclear if we need to use the _ModelCache class in the same way that apps/units do. It seems like _ModelCache serves a different purpose than caching the results of a backend call (caches an OF object rather than the Juju call results).