remove intermediate assertions using model.goal and add cached goal-state test
Since the calls to goal-state are being cached, we need to only make assertions of model.goal.num_units and model.goal.pending_units at the end of a test.

This commit also adds a test that checks if the result of goal-state was cached by checking that setting (previously pending) unit statuses to active does not cause a change in model.goal.pending_units.