Low-level Pebble API client (#468)
A `PebbleClient` that speaks the HTTP-over-UNIX-socket protocol that the Pebble server uses. Unit tests for the above (I'm working on more full-fledged tests using a fake Pebble client that actually talks HTTP, but that'll come in a separate PR).

Note: we're going to rename `pebble.API` to `pebble.PebbleClient` in a subsequent PR, per discussion with Gustavo.