Support time offset 'Z' (UTC) and no subseconds in _parse_timestamp (#476)
Z for +00:00 (UTC) is as per RFC3339 and also what Go produces. I had just missed it before, because I was testing on NZ time (+13:00). Add a unit test for this, and also a few for failure cases.

Also change _parse_timestamp to convert directly, and handle no fractional seconds.

Might as well support lower case 't' and 'z' as RFC3339 allows them.

Use round() instead of truncating when converting to microseconds, to get closer.