Dynamic add_layer API and get_plan API (#483)
This adds `add_layer` and `get_plan` to both the `Pebble` instance as well as `model.Container`. This is the Python / Operator Framework end of https://github.com/canonical/pebble/pull/10

Note that these changes would be a breaking change, but the functions in question currently raise `NotImplementedError`, and nobody except me is using it yet anyway (we haven't made an Operator Framework release with this in them at all).