On Windows, NamedTemporaryFile should be explicitly closed. (#504)
If you don't explicitly call close() you end up with warnings about
leaked file descriptors.

On non-NT the NamedTemporaryFile wrapper has:

```
        def __del__(self):
            self.close()
```

However, on nt it has a 'delete-on-close' primitive so it doesn't try to
force call close to unlink the file. However, that means close() isn't
called on teardown either, which means the file descriptor leaks.