Pebble testing (#515)
Fixes #488

This implements a fairly primitive pebble backend for the testing Harness. It allows you to publish layers and read what the overall plan is. It also allows you to start and stop services and check whether they are running.

It only supports `override: replace` directives (doesn't support `override:merge`). But for the straightforward cases of asking pebble to track a layer and start some services, you have the opportunity to set that up with the testing Harness, and assert that you have the right plan when you are done. And that the charm has asked for the service to start.