Added support to remove specific relation by ID
Multiple relations with the same name can exist between the same
two charms/applications. This commit adds support to remove a
specific relation in such a case. The relation removed is identified
by its relation ID. This ID is an optional argument to the
remove relation harness method.

This commit also adds a unit test to validate the changed behaviour
of the remove relation method.