A pair of Provider and Consumer utility objects (#502)
* A pair of Provider and Consumer utility objects

This commit adds a utility module that handles relation events between
resource providers and consumers. The Provider and Consumer objects
check type and version compatibilty between the resource provider and
resource consumer.

* Consumer now raises TooManyProviders

This commit adds a new event that is received by Consumer charms.
This TooManyProviders event is emitted when more than one provider
charm unit are related to a consumer charm unit that only accepts
one provider.

* A property to access stored relation id in consumer

This commit provides a property to access the relation id
of ConsumerBase. This is to discourage access of
StoreState across class boundaries for reasons described in
https://github.com/canonical/operator/issues/526 .

* Added semantic version package as a dependency

This commit adds the semantic version package as a dependench since
the provider consumer utility depends on it.