Expose 'limit' in RelationMeta and improve ‘scope’ (#552)
his PR adds support to limit in the RelationMeta class, which is currently not accessible at runtime by charms, and adds type validation and correctly sets the default value for RelationMeta.scope.

The metadata.yaml spec allows to limit the cardinality of relations using the limit keyword, e.g., when we want up to one relation in the model using the mongodb-datasource interface. Additionally, the enumeration of the possible values for RelationMeta.scope (namely: global, which is default, or container) are not validated, and unless the scope is explicitly set, RelationMeta.scope is set to None rather than the default global.