Support tests for removed relations and units (#460)
* Remove duplicate leader initialization

There were two is_leader member variable initializers in the
test harness backend constructor. This commit removes one of them.

* Added remove relation and unit methods to harness

This commit adds method to remove a relation and a
relation unit to the test harness.

* Added tests for remove relation and unit

This commit adds unit tests for the test harness method that
allows removing relation and relation units.

* Added support to remove specific relation by ID

Multiple relations with the same name can exist between the same
two charms/applications. This commit adds support to remove a
specific relation in such a case. The relation removed is identified
by its relation ID. This ID is an optional argument to the
remove relation harness method.

This commit also adds a unit test to validate the changed behaviour
of the remove relation method.

Co-authored-by: Ryan Barry <ryan.barry@canonical.com>