Check for ready in more places, add `Container.restart()`
Since there are almost no operations we can actually perform if
Pebble returns a `ConnectionError`, check for it first and log
it if that's the case to give a hint to Charm authors in case they
forget to check.

This required some extra test mocking due to the way
`MockPebbleClient` is implemented.

Add a `restart` method to `ops.model.Container` so authors do not
need to write their own methods for it